updated file structure, ran all tests, looking good to go
