+++ benchmarks/binaryHeap.spec.js
@@ -4,7 +4,7 @@
 
 'use strict';
 
-var Needle = require('../src/needle.js');
+var Needle = require('../src');
 var winston = require('winston');
 winston.log('info', 'running speed tests on binary heap');
 var start, end, time, iterations, i;
