+++ src/BinaryHeap/index.js
@@ -1,0 +1,5 @@
+'use strict';
+
+const BinaryHeap = require('./binaryHeap');
+
+module.exports = BinaryHeap;
