+++ src/BinarySearchTree/index.js
@@ -1,0 +1,5 @@
+'use strict';
+
+const BinarySearchTree = require('./binarySearchTree');
+
+module.exports = BinarySearchTree;
