+++ src/BitArray/bitArray.js
