+++ src/BitArray/index.js
@@ -1,0 +1,5 @@
+'use strict';
+
+const BitArray = require('./bitArray');
+
+module.exports = BitArray;
