+++ src/DoublyLinkedList/index.js
@@ -1,0 +1,5 @@
+'use strict';
+
+const DoublyLinkedList = require('./doublyLinkedList');
+
+module.exports = DoublyLinkedList;
