+++ src/Hashmap/index.js
@@ -1,0 +1,5 @@
+'use strict';
+
+const Hashmap = require('./hashmap');
+
+module.exports = Hashmap;
