+++ src/KaryTree/index.js
@@ -1,0 +1,5 @@
+'use strict';
+
+const KaryTree = require('./KaryTree');
+
+module.exports = KaryTree;
