+++ src/Queue/index.js
@@ -1,0 +1,5 @@
+'use strict';
+
+const Queue = require('./queue');
+
+module.exports = Queue;
