+++ src/Queue/queue.js
@@ -12,7 +12,7 @@
  * 
  */
 
-const Node = require('./nodes/unidirectional_node.js');
+const Node = require('../__Nodes__/unidirectional_node.js');
 
 /**
  * Single argument constructor.
