+++ src/RollingHash/index.js
@@ -1,0 +1,5 @@
+'use strict';
+
+const RollingHash = require('./rollingHash');
+
+module.exports = RollingHash;
