+++ src/RollingHash/rollingHash.js
