+++ src/SinglyLinkedList/index.js
@@ -1,0 +1,5 @@
+'use strict';
+
+const SinglyLinkedList = require('./singlyLinkedList');
+
+module.exports = SinglyLinkedList;
