+++ src/SinglyLinkedList/singlyLinkedList.js
@@ -1,7 +1,6 @@
 /**
- * Doubly Linked List
+ * Singly Linked List
  * {head} Node, the head of the linked list
- * {tail} Node, the end of the linked list
  * {size} number, the number of nodes in the linked list
  * 
  * Asymptotic time complexities
@@ -20,24 +19,21 @@
  
 'use strict';
 
-const Node = require('./nodes/bidirectional_node.js');
+const Node = require('../__Nodes__/unidirectional_node.js');
 
 /**
  * Single argument constructor.
  * @param {*} [data] data for head node of linked list
  * @return {void}
  */
-const DoublyLinkedList = function(data){
+const SinglyLinkedList = function(data){
   this.head;
-  this.tail;
   this.size;
   if(typeof data !== 'undefined'){
     this.head = new Node(data);
-    this.tail = this.head;
     this.size = 1;
   }else{
     this.head = null;
-    this.tail = null;
     this.size = 0;
   }
 }
@@ -48,29 +44,22 @@ const DoublyLinkedList = function(data){
  * @param {*} data for the new node
  * @return {void}
  */
-DoublyLinkedList.prototype.insertFront = function(data){
+SinglyLinkedList.prototype.insertFront = function(data){
   if(typeof data === 'undefined'){
-    throw new Error("Too few arguments for DoublyLinkedList.insertFront");
+    throw new Error("Too few arguments for SinglyLinkedList.insertFront");
   }
 
   var nodeToAdd = new Node(data);
   
   // If the head has not been set, add this node as the head 
-  if(this.head === null && this.tail === null){
+  if(this.head === null){
     this.head = nodeToAdd;
-    this.tail = this.head;
   }
-  // If head and tail are set, add node to front of the list
-  else if(this.head !== null && this.tail !== null){
+  // If head is set, add node to front of the list
+  else{
     var prevHead = this.head;
-    prevHead.prev = nodeToAdd;
     this.head = nodeToAdd;
     this.head.next = prevHead;
-    this.head.prev = null;
-  }
-  // If head and tail are not in sync, throw error
-  else{
-    throw new Error("Unsynced head and tail in an Doubly Linked List. Please report this to https://github.com/nickzuber/needle/issues");
   }
   ++this.size;
 }
@@ -82,17 +71,17 @@ DoublyLinkedList.prototype.insertFront = function(data){
  * @param {*} data for the new node
  * @return {void}
  */
-DoublyLinkedList.prototype.insertNth = function(index, data){
+SinglyLinkedList.prototype.insertNth = function(index, data){
   if(typeof data === 'undefined' || typeof index === 'undefined'){
-    throw new Error("Too few arguments for DoublyLinkedList.insertNth");
+    throw new Error("Too few arguments for SinglyLinkedList.insertNth");
   }else if(typeof index !== 'number'){
-    throw new TypeError("Invalid argument for DoublyLinkedList.insertNth");
+    throw new TypeError("Invalid argument for SinglyLinkedList.insertNth");
   }
   // Check for bounds (and if inserting at index 0 we want to allow)
   else if(index < 0 || index >= this.size && index !== 0){
-    throw new Error("Index out of bounds on DoublyLinkedList.insertNth");
+    throw new Error("Index out of bounds on SinglyLinkedList.insertNth");
   }
-  
+
   // Check if inserting at head
   if(index === 0){
     this.insertFront(data);
@@ -104,14 +93,7 @@ DoublyLinkedList.prototype.insertNth = function(index, data){
     return false;
   }
 
-  // Check if inserting at tail
-  if(index === this.size-1){
-    this.insertBack(data);
-    return;
-  }
-
   // Configure finder nodes
-  // We can assume we are not inserting at head
   var nodeToAdd = new Node(data),
       curNode = this.head;
 
@@ -121,22 +103,19 @@ DoublyLinkedList.prototype.insertNth = function(index, data){
 
   var tempNode = curNode.next;
   curNode.next = nodeToAdd;
-  tempNode.prev = nodeToAdd;
-  nodeToAdd.prev = curNode;
   nodeToAdd.next = tempNode;
   ++this.size;
 }
 
 /**
  * Create a node from given data and inserts after a given node.
- * of linked list.
  * @param {*} the data of the node which the new node is to be inserted after
  * @param {*} data for the new node
  * @return {boolean} returns if node is inserted or not
  */
-DoublyLinkedList.prototype.insertAfter = function(targetData, data){
+SinglyLinkedList.prototype.insertAfter = function(targetData, data){
   if(typeof data === 'undefined' || typeof targetData === 'undefined'){
-    throw new Error("Too few arguments for DoublyLinkedList.insertAfter");
+    throw new Error("Too few arguments for SinglyLinkedList.insertAfter");
   }
   // If head is empty, we obviously won't find the targetNode
   if(this.head === null){
@@ -156,17 +135,6 @@ DoublyLinkedList.prototype.insertAfter = function(targetData, data){
   }else{
     _targetData = targetData;
   }
-  if(typeof curNode.data === 'object'){
-    _curNodedata = JSON.stringify(curNode.data);
-  }else{
-    _curNodedata = curNode.data;
-  }
-
-  // Check if inserting after tail
-  if(JSON.stringify(this.tail.data) === JSON.stringify(targetData)){
-    this.insertBack(data);
-    return true;
-  }
 
   while(curNode !== null){
     // JSON.stringify is very expensive - only perform if necessary
@@ -185,8 +153,6 @@ DoublyLinkedList.prototype.insertAfter = function(targetData, data){
   if(nodeFound){
     var tempNode = curNode.next;
     curNode.next = nodeToAdd;
-    tempNode.prev = nodeToAdd;
-    nodeToAdd.prev = curNode;
     nodeToAdd.next = tempNode;
     ++this.size;
     return true;
@@ -201,9 +167,9 @@ DoublyLinkedList.prototype.insertAfter = function(targetData, data){
  * @param {*} data for the new node
  * @return {void}
  */
-DoublyLinkedList.prototype.insertBack = function(data){
+SinglyLinkedList.prototype.insertBack = function(data){
   if(typeof data === 'undefined'){
-    throw new Error("Too few arguments for DoublyLinkedList.insert");
+    throw new Error("Too few arguments for SinglyLinkedList.insertBack");
   }
 
   var nodeToAdd = new Node(data);
@@ -211,14 +177,15 @@ DoublyLinkedList.prototype.insertBack = function(data){
   // If the head has not been set, add this node as the head 
   if(this.head === null){
     this.head = nodeToAdd;
-    this.tail = this.head;
   }
   // If head is set, add node to end of the list
   else{
-    nodeToAdd.next = null;
-    nodeToAdd.prev = this.tail;
-    this.tail.next = nodeToAdd;
-    this.tail = nodeToAdd;
+    var curNode = this.head;
+    // Find the node without a next element (will be the tail)
+    while(curNode.next !== null){
+      curNode = curNode.next;
+    }
+    curNode.next = nodeToAdd;
   }
   ++this.size;
 }
@@ -229,16 +196,15 @@ DoublyLinkedList.prototype.insertBack = function(data){
  * @return {boolean} true if node removal was successful
  *                   false if node not found
  */
-DoublyLinkedList.prototype.remove = function(data){
+SinglyLinkedList.prototype.remove = function(data){
   if(typeof data === 'undefined'){
-    throw new Error("Too few arguments for DoublyLinkedList.remove");
+    throw new Error("Too few arguments for SinglyLinkedList.remove");
   }
   else if(this.size <= 0){
-    throw new Error("Attempted to remove from an empty DoublyLinkedList");
+    throw new Error("Attempted to remove from an empty SinglyLinkedList");
   }
   else if(this.size === 1){
     this.head = null;
-    this.tail = null;
     this.size = 0;
     return;
   }
@@ -247,7 +213,8 @@ DoublyLinkedList.prototype.remove = function(data){
   }
 
   // Configure finder nodes
-  var curNode = this.head,
+  var prevNode = null,
+      curNode = this.head,
       nodeFound = false,
       _curNodedata,
       _data;
@@ -265,28 +232,19 @@ DoublyLinkedList.prototype.remove = function(data){
   }
 
   // Check if deleting head
-  if(JSON.stringify(this.head.data) === JSON.stringify(data)){
+  if(_curNodedata === _data){
     this.head = curNode.next;
-    this.head.prev = null;
-    curNode = null;
-    --this.size;
-    return true;
-  }
-
-  // Check if deleting tail
-  if(JSON.stringify(this.tail.data) === JSON.stringify(data)){
-    curNode = this.tail;
-    this.tail = curNode.prev;
-    this.tail.next = null;
     curNode = null;
     --this.size;
     return true;
   }
 
   // Update nodes to account for head not being the node to delete
-  curNode = curNode.next;
+  prevNode = curNode;
+  curNode = prevNode.next;
 
   while(curNode !== null){
+    // JSON.stringify is very expensive - only perform if necessary
     if(typeof curNode.data === 'object'){
       _curNodedata = JSON.stringify(curNode.data);
     }else{
@@ -294,13 +252,12 @@ DoublyLinkedList.prototype.remove = function(data){
     }
     if(_curNodedata === _data){
       nodeFound = true;
-      var prevNode = curNode.prev;
       prevNode.next = curNode.next;
-      curNode.next.prev = prevNode;
       curNode = null;
       --this.size;
       break;
     }
+    prevNode = curNode;
     curNode = curNode.next;
   }
 
@@ -313,54 +270,43 @@ DoublyLinkedList.prototype.remove = function(data){
  * @param {number} index of node to remove
  * @return {void}
  */
-DoublyLinkedList.prototype.removeNth = function(index){
+SinglyLinkedList.prototype.removeNth = function(index){
   if(typeof index === 'undefined'){
-    throw new Error("Too few arguments for DoublyLinkedList.removeNth");
+    throw new Error("Too few arguments for SinglyLinkedList.removeNth");
   }
   else if(typeof index !== 'number'){
-    throw new TypeError("Invalid argument for DoublyLinkedList.removeNth");
+    throw new TypeError("Invalid argument for SinglyLinkedList.removeNth");
   }
   // Check for bounds
   else if(index < 0 || index >= this.size){
-    throw new Error("Index out of bounds on DoublyLinkedList.removeNth: " + index);
+    throw new Error("Index out of bounds on SinglyLinkedList.removeNth: " + index);
   }
   else if(this.size === 1){
     this.head = null;
-    this.tail = null;
     this.size = 0;
     return;
   }
 
+  var prevNode = null;
   var curNode = this.head;
   
   // Check if deleting head
   if(index === 0){
     this.head = curNode.next;
-    this.head.prev = null;
     curNode = null;
     --this.size;
     return;
   }
 
-  // Check if deleting tail
-  if(index === this.size-1){
-    curNode = this.tail;
-    this.tail = curNode.prev;
-    this.tail.next = null;
-    curNode = null;
-    --this.size;
-    return true;
-  }
-
+  prevNode = curNode;
   curNode = curNode.next;
 
   for(var i=1; i<index; ++i){
+    prevNode = curNode;
     curNode = curNode.next;
   }
 
-  var prevNode = curNode.prev;
   prevNode.next = curNode.next;
-  curNode.next.prev = prevNode;
   curNode = null;
   --this.size;
   return;
@@ -372,12 +318,12 @@ DoublyLinkedList.prototype.removeNth = function(index){
  * @param {*} data of node to find
  * @return {boolean || Node} returns node found and false if not found
  */
-DoublyLinkedList.prototype.find = function(data){
+SinglyLinkedList.prototype.find = function(data){
   if(typeof data === 'undefined'){
-    throw new Error("Too few arguments for DoublyLinkedList.find");
+    throw new Error("Too few arguments for SinglyLinkedList.find");
   }
   else if(this.size <= 0){
-    throw new Error("Attempted to find a node from an empty DoublyLinkedList");
+    throw new Error("Attempted to find a node from an empty SinglyLinkedList");
   }
   else if(this.head === null && this.size > 0){
     throw new Error("Null head in an unemptied list. Please report this to https://github.com/nickzuber/needle/issues");
@@ -407,11 +353,11 @@ DoublyLinkedList.prototype.find = function(data){
  * @param {number} index of node to find
  * @returm {Node} the node at the given position
  */
-DoublyLinkedList.prototype.findNth = function(index){
+SinglyLinkedList.prototype.findNth = function(index){
   if(typeof index === 'undefined'){
-    throw new TypeError("Too few arguments for DoublyLinkedList.findNth");
+    throw new TypeError("Too few arguments for SinglyLinkedList.findNth");
   }else if(typeof index !== 'number'){
-    throw new TypeError("Invalid argument for DoublyLinkedList.findNth");
+    throw new TypeError("Invalid argument for SinglyLinkedList.findNth");
   }
   // Check for bounds
   else if(index < 0 || index >= this.size){
@@ -430,4 +376,4 @@ DoublyLinkedList.prototype.findNth = function(index){
   return curNode;
 }
 
-module.exports = DoublyLinkedList;
+module.exports = SinglyLinkedList;
