+++ src/SortedArray/index.js
@@ -1,0 +1,5 @@
+'use strict';
+
+const SortedArray = require('./sortedArray');
+
+module.exports = SortedArray;
