+++ src/Stack/index.js
@@ -1,0 +1,5 @@
+'use strict';
+
+const Stack = require('./stack');
+
+module.exports = Stack;
