+++ src/Stack/stack.js
@@ -14,7 +14,7 @@
 
 'use strict';
 
-const Node = require('./nodes/unidirectional_node.js');
+const Node = require('../__Nodes__/unidirectional_node.js');
 
 /**
  * Single argument constructor.
