+++ src/Trie/index.js
@@ -1,0 +1,5 @@
+'use strict';
+
+const Trie = require('./trie');
+
+module.exports = Trie;
