+++ src/__Nodes__/bidirectional_node.js
