+++ src/__Nodes__/bidirectional_tree_node.js
