+++ src/__Nodes__/multidirectional_tree_node.js
