+++ src/__Nodes__/unidirectional_node.js
