+++ unit/README.md
@@ -1,3 +1,9 @@
 # Unit Testing <img src="https://travis-ci.org/nickzuber/needle.svg?branch=master" /> <img src="https://badge.fury.io/js/node-needle.svg" /> <img src="https://david-dm.org/nickzuber/Needle.svg" /> <img src="https://david-dm.org/nickzuber/Needle/dev-status.svg" />
 
 The files located in this folder are used to test the functionality and behavior of each data structure which is implemented in Needle.
+
+From the root:
+
+```
+$ npm run unit
+```
