+++ unit/binarySearchTree.unit.js
@@ -1,7 +1,7 @@
 
 // # Binary Search Tree tests
 
-var Needle = require('../src/needle.js');
+var Needle = require('../src');
 var test = require('tape');
 
 test('Binary Search Tree should run without errors', function(t){
