+++ unit/doublyLinkedList.unit.js
@@ -1,7 +1,7 @@
 
 // # Doubly Linked List tests
 
-var Needle = require('../src/needle.js');
+var Needle = require('../src');
 var test = require('tape');
 
 test('Doubly Linked List should run without errors', function(t){
