+++ unit/rollingHash.unit.js
@@ -1,7 +1,7 @@
 
 // # Hashmap tests
 
-var Needle = require('../src/needle.js');
+var Needle = require('../src');
 var test = require('tape');
 
 test('Rolling Hash should run without errors', function(t){
