+++ unit/singlyLinkedList.unit.js
@@ -1,7 +1,7 @@
 
 // # Singly Linked List tests
 
-var Needle = require('../src/needle.js');
+var Needle = require('../src');
 var test = require('tape');
 
 test('Singly Linked List should run without errors', function(t){
