+++ unit/sortedArray.unit.js
@@ -1,7 +1,7 @@
 
 // # SortedArray tests
 
-var Needle = require('../src/needle.js');
+var Needle = require('../src');
 var test = require('tape');
 
 test('SortedArray should run without errors', function(t){
